requirements.txt
# Merlai Requirements
# Generated from current development environment
# Install with: pip install -r requirements.txt

accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
decorator==5.2.1
diffusers==0.33.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
execnet==2.1.1
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.3.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
MIDIUtil==1.2.1
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.1
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.2.0
pretty_midi==0.2.10
psutil==7.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
types-PyYAML
